requirements.txt
django
djangorestframework
drf-yasg
python-decouple
psycopg2-binary
djangorestframework-simplejwt
django-rest-framework-social-oauth2
django-filter
Pillow